Reorganize tests in to multiple files so the tests are less overwhelming to work with
