+++ tests/smoke-tests/--test-app-location.test.ts
@@ -1,0 +1,45 @@
+import fs from 'node:fs/promises';
+import path from 'node:path';
+import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+
+import { assertGeneratedCorrectly } from '../assertions.js';
+import { createAddon, createTmp, install, runScript } from '../utils.js';
+
+describe('--test-app-location', () => {
+  let cwd = '';
+  let tmpDir = '';
+  let testAppLocation = 'packages/my-custom-location';
+
+  beforeAll(async () => {
+    tmpDir = await createTmp();
+
+    let { name } = await createAddon({
+      args: [`--test-app-location=${testAppLocation}`, '--pnpm=true'],
+      options: { cwd: tmpDir },
+    });
+
+    cwd = path.join(tmpDir, name);
+
+    await install({ cwd, packageManager: 'pnpm' });
+  });
+
+  afterAll(async () => {
+    fs.rm(tmpDir, { recursive: true, force: true });
+  });
+
+  it('was generated correctly', async () => {
+    assertGeneratedCorrectly({ projectRoot: cwd, testAppLocation });
+  });
+
+  it('runs tests', async () => {
+    let { exitCode } = await runScript({ cwd, script: 'test', packageManager: 'pnpm' });
+
+    expect(exitCode).toEqual(0);
+  });
+
+  it('lints all pass', async () => {
+    let { exitCode } = await runScript({ cwd, script: 'lint', packageManager: 'pnpm' });
+
+    expect(exitCode).toEqual(0);
+  });
+});
